Move manifest funcs
